Title: 5와 7의 배수를 제외하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240918/5와 7의 배수를 제외하기/excluding-multiples-of-5-and-7.cpp
@@ -7,7 +7,7 @@ int main() {
     int n, input;
     float sum = 0;
     float avg = 0;
-    float cnt = 0;
+    int cnt = 0;
 
     cin >> n;
